+++ spinetoolbox/widgets/settings_widget.py
@@ -283,7 +283,7 @@ class SettingsWidget(QWidget):
         else:  # commit_at_exit == "2":
             self.ui.checkBox_commit_at_exit.setCheckState(Qt.Checked)
         if sticky_selection == "true":
-            self.ui.checkBox_graph_view_sticky_selection.setCheckState(Qt.Checked)
+            self.ui.checkBox_object_tree_sticky_selection.setCheckState(Qt.Checked)
         proj_dir = ""  # Unused. Save/read this using QSettings, if we want to change the projects dir at some point
         if not proj_dir:
             proj_dir = DEFAULT_PROJECT_DIR
@@ -363,7 +363,7 @@ class SettingsWidget(QWidget):
         # Data Store Views
         commit_at_exit = str(int(self.ui.checkBox_commit_at_exit.checkState()))
         self._qsettings.setValue("appSettings/commitAtExit", commit_at_exit)
-        sticky_selection = "true" if int(self.ui.checkBox_graph_view_sticky_selection.checkState()) else "false"
+        sticky_selection = "true" if int(self.ui.checkBox_object_tree_sticky_selection.checkState()) else "false"
         self._qsettings.setValue("appSettings/stickySelection", sticky_selection)
         # Check if something in the app needs to be updated
         self._toolbox.show_datetime = self._toolbox.update_datetime()